requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
beautifulsoup4>=4.12.2
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=7.4.0